Refactoring utils directory with barcode, cryptography, frontend, solid
