+++ src/utils/barcode/barcode-scan.js
